[REFACTOR/#64] AuthInterceptor data 모듈로 이전
