+++ app/src/main/java/com/going/doorip/di/RetrofitModule.kt
@@ -1,5 +1,6 @@
 package com.going.doorip.di
 
+import com.going.data.interceptor.AuthInterceptor
 import com.going.doorip.BuildConfig.BASE_URL
 import com.going.doorip.di.qualifier.JWT
 import com.going.ui.extension.isJsonArray
