+++ data/src/main/java/com/going/data/interceptor/AuthInterceptor.kt
@@ -1,9 +1,9 @@
-package com.going.doorip.di
+package com.going.data.interceptor
 
+import com.going.data.BuildConfig.BASE_URL
 import com.going.data.dto.BaseResponse
+import com.going.data.local.GoingDataStore
 import com.going.domain.entity.response.AuthTokenModel
-import com.going.domain.repository.TokenRepository
-import com.going.doorip.BuildConfig.BASE_URL
 import kotlinx.serialization.json.Json
 import okhttp3.Interceptor
 import okhttp3.Request
@@ -14,17 +14,15 @@ import javax.inject.Inject
 
 class AuthInterceptor @Inject constructor(
     private val json: Json,
-    private val tokenRepository: TokenRepository,
+    private val dataStore: GoingDataStore,
 ) : Interceptor {
 
     override fun intercept(chain: Interceptor.Chain): Response {
         val originalRequest = chain.request()
 
-        Timber.d("GET ACCESS TOKEN : ${tokenRepository.getAccessToken()}")
+        Timber.d("GET ACCESS TOKEN : ${dataStore.accessToken}")
 
-        val authRequest = if (tokenRepository.getAccessToken()
-                .isNotBlank()
-        ) {
+        val authRequest = if (dataStore.accessToken.isNotBlank()) {
             originalRequest.newAuthBuilder().build()
         } else {
             originalRequest
@@ -35,15 +33,15 @@ class AuthInterceptor @Inject constructor(
             CODE_TOKEN_EXPIRED -> {
                 try {
                     // refreshToken을 얻기 위해서 인터셉터 내부에서 직접적으로 통신을 하는게 맞을까?
-                    // app 레이어에서 data layer에 직접 접근하는 것도 맘에 안들고, 직접 통신하는게 아키텍쳐 구조를 헤친다는 생각이 듦
-                    // 속도를 측정해보고 repository를 사용한 통신방법이 속도가 느리지 않다면
-                    // 추후 repository를 활용한 통신으로 수정할 예정
+                    // 직접 통신하는게 기존의 통신방법과 차이가 있기 때문에 통일성을 해친다는 생각이 듦
+                    // 그리고 이렇게 되면 Base Url이 App모듈에도 필요하고, Data모듈에도 필요함
+                    // 그래서 지금 생각으론 Retrofit 관련 모듈이 data 모듈에 있어야 할 것 같음
                     // 앱잼 기간내에 수정하는게 목표라 주석을 남겨두겠습니다. 수정 못할거 같으면 주석만 지울게용
                     val refreshTokenRequest = originalRequest.newBuilder().post("".toRequestBody())
                         .url("$BASE_URL/api/users/reissue")
                         .addHeader(
                             AUTHORIZATION,
-                            tokenRepository.getRefreshToken(),
+                            dataStore.refreshToken,
                         )
                         .build()
                     val refreshTokenResponse = chain.proceed(refreshTokenRequest)
@@ -54,21 +52,30 @@ class AuthInterceptor @Inject constructor(
                             refreshTokenResponse.body?.string().toString(),
                         ) as BaseResponse<AuthTokenModel>
 
-                        tokenRepository.setTokens(
-                            accessToken = responseToken.data.accessToken,
-                            refreshToken = responseToken.data.refreshToken,
-                        )
+                        dataStore.apply {
+                            accessToken = responseToken.data.accessToken
+                            refreshToken = responseToken.data.refreshToken
+                        }
+
                         refreshTokenResponse.close()
+
                         val newRequest = originalRequest.newAuthBuilder().build()
                         return chain.proceed(newRequest)
                     }
 
-                    tokenRepository.clearTokens()
+                    dataStore.apply {
+                        accessToken = ""
+                        refreshToken = ""
+                    }
 
                     // refreshToken 만료 처리를 위한 리프레시 토큰 만료 코드 포함 리스폰스 리턴
                     return refreshTokenResponse
                 } catch (t: Throwable) {
-                    tokenRepository.clearTokens()
+                    dataStore.apply {
+                        accessToken = ""
+                        refreshToken = ""
+                    }
+
                     Timber.e(t)
                 }
             }
@@ -77,7 +84,7 @@ class AuthInterceptor @Inject constructor(
     }
 
     private fun Request.newAuthBuilder() =
-        this.newBuilder().addHeader(AUTHORIZATION, "$BEARER ${tokenRepository.getAccessToken()}")
+        this.newBuilder().addHeader(AUTHORIZATION, "$BEARER ${dataStore.accessToken}")
 
     companion object {
         private const val CODE_TOKEN_EXPIRED = 401
